+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **Bugs**
 - Shadow for delete and edit card buttons
 - Soltan palette color fix
+- Theme selector fix
 
 # Version 1.8.0 (2025-06-20)